#35: Remove code using the core API
