--- agent/src/main/kotlin/com/octogonapus/omj/agent/OMJMethodAdapter.kt
@@ -1,134 +1,0 @@
-/*
- * This file is part of OMJ.
- *
- * OMJ is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * OMJ is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with OMJ.  If not, see <https://www.gnu.org/licenses/>.
- */
-package com.octogonapus.omj.agent
-
-import com.octogonapus.omj.di.OMJKoinComponent
-import mu.KotlinLogging
-import org.koin.core.get
-import org.koin.core.inject
-import org.koin.core.qualifier.named
-import org.objectweb.asm.Label
-import org.objectweb.asm.MethodVisitor
-import org.objectweb.asm.Opcodes.GETFIELD
-import org.objectweb.asm.Opcodes.GETSTATIC
-import org.objectweb.asm.Opcodes.PUTFIELD
-import org.objectweb.asm.Opcodes.PUTSTATIC
-
-internal class OMJMethodAdapter(
-    api: Int,
-    private val superVisitor: MethodVisitor,
-    private val method: Method,
-    private val isStatic: Boolean,
-    currentClassName: String
-) : MethodVisitor(api, superVisitor), OMJKoinComponent {
-
-    private val dynamicClassDefiner by inject<DynamicClassDefiner>()
-    private val classFilter by inject<ClassFilter>()
-    private val methodAdapterUtil by inject<MethodAdapterUtil>()
-    private val fullyQualifiedClassName =
-        MethodAdapterUtil.convertPathTypeToPackageType(currentClassName)
-    private var currentLineNumber = 0
-
-    override fun visitCode() {
-        super.visitCode()
-        methodAdapterUtil.visitMethodTrace(
-            superVisitor,
-            method.descriptor,
-            isStatic,
-            dynamicClassDefiner
-        )
-    }
-
-    override fun visitMethodInsn(
-        opcode: Int,
-        owner: String,
-        name: String,
-        descriptor: String,
-        isInterface: Boolean
-    ) {
-        @Suppress("DuplicatedCode")
-        logger.debug {
-            """
-            opcode = $opcode
-            owner = $owner
-            name = $name
-            descriptor = $descriptor
-            isInterface = $isInterface
-            """.trimIndent()
-        }
-
-        // Only add the preamble to methods which we will also record a trace for
-        if (classFilter.shouldTransform(owner)) {
-            methodAdapterUtil.visitMethodCallStartPreamble(
-                superVisitor,
-                currentLineNumber,
-                fullyQualifiedClassName,
-                name
-            )
-        }
-
-        super.visitMethodInsn(opcode, owner, name, descriptor, isInterface)
-    }
-
-    override fun visitVarInsn(opcode: Int, index: Int) {
-        methodAdapterUtil.visitVarInsn(
-            superVisitor,
-            fullyQualifiedClassName,
-            currentLineNumber,
-            opcode,
-            index,
-            get<MethodsAndLocals>(named(METHODS_AND_LOCALS_NAME))[method]
-        )
-    }
-
-    override fun visitIincInsn(index: Int, increment: Int) {
-        methodAdapterUtil.visitIincInsn(
-            superVisitor,
-            fullyQualifiedClassName,
-            currentLineNumber,
-            index,
-            increment,
-            get<MethodsAndLocals>(named(METHODS_AND_LOCALS_NAME))[method]
-        )
-    }
-
-    override fun visitFieldInsn(opcode: Int, owner: String, name: String, descriptor: String) {
-        when (opcode) {
-            PUTFIELD, PUTSTATIC -> methodAdapterUtil.visitFieldInsn(
-                superVisitor,
-                fullyQualifiedClassName,
-                currentLineNumber,
-                opcode,
-                owner,
-                name,
-                descriptor
-            )
-
-            GETFIELD, GETSTATIC -> super.visitFieldInsn(opcode, owner, name, descriptor)
-        }
-    }
-
-    override fun visitLineNumber(line: Int, start: Label) {
-        super.visitLineNumber(line, start)
-        currentLineNumber = line
-    }
-
-    companion object {
-
-        private val logger = KotlinLogging.logger { }
-    }
-}
